% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_gen_tbl}
\alias{a_gen_tbl}
\title{Generate Information Table}
\usage{
a_gen_tbl(pat_sub, stack_by, distn)
}
\arguments{
\item{pat_sub}{A data set contains needed information for ploting}

\item{stack_by}{A string indicating the way to group data. Options are "sex" and "diagnosis".}

\item{distn}{A string indicating the interested distribution. Options are "age", "visit", and "yar"}
}
\value{
A matrix with number of participants in different categories.
}
\description{
Generate Information Table
}
\examples{
\dontrun{
a_gen_tbl('sex', 'age')
}
}
