% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_gsub}
\alias{a_gsub}
\title{Remove all special characters}
\usage{
a_gsub(vec_name, lower_case = TRUE)
}
\arguments{
\item{vec_name}{A character string indicating the vector needed to be removed special characters.}

\item{lower_case}{A logical value indicating the lower case setting.
Default "False." If true, all characters will to convert to lower cases.}
}
\value{
Returned a string with no special characters except "_"
}
\description{
Remove all special charachers in a string except "_",
and change all characters to the lower cases is set lower_case = TRUE.
}
\examples{
\dontrun{
vec <- "ad*&_ae20"
a_gsub(vec, lower_case = TRUE)
}


}
