% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_map_var}
\alias{a_map_var}
\title{Get the column name in the source data file.}
\usage{
a_map_var(
  src_type = c("BIOCARD", "NACC", "ADNI"),
  tbl_code,
  col_name,
  dict_src_tables = NULL
)
}
\arguments{
\item{src_type}{A string indicating the source of data. Options include
"BIOCARD", "NACC" and "ADNI"}
}
\value{
The real name in the source data file
}
\description{
Since the column names in tables may vary, it could be modified by using a
costomized column names dictionary. The format can refer to the appended
dictionary.
}
