% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_match}
\alias{a_match}
\title{Map Biomarkers Data to The Baseline Visits}
\usage{
a_match(dat_se, dat_marker, m_date)
}
\arguments{
\item{dat_se}{A dataset that will be merged.}

\item{dat_marker}{A data set including new biomarker that will be merged.}

\item{m_date}{A string indicating the name of measure date in dat_marker.}
}
\value{
The matched (larger) dataset including newly added biomarker data.
}
\description{
Map all marker data to baseline visits. For each visit, assign the
closest markers within the setted window. If no marker
satisfies assignment criteria, mark as missing. If two marker satisfy
assignment criteria, take first one.
}
\examples{
\dontrun{
dat_se <- a_match(dat_se, dat_dx, "date_dx")
}

}
