% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_window}
\alias{a_window}
\title{Define the Time Window}
\usage{
a_window(dat, v_date, window, window_overlap, v_id = "subject_id")
}
\arguments{
\item{dat}{The baseline time dataset.}

\item{v_date}{A string indicating the date name of the baseline time.}

\item{v_id}{A string indicating the id name of baseline dateset. Default is
"subject_id" (may varied if using customized column names dictionary).}
}
\description{
For most data, the time windows are calculated from the baseline time (left
window = the midpoint between current and the previous time point; right
window = the midpoint between current and the next time point). If the time
windows are longer than the maximum acceptable length, then force to select
biomarker within the maximum window length we set. For the first and last
baseline time, since there is no "previous" or "next" time point availiable,
use the maximum acceptable window length instead.
}
