% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_apoe}
\alias{adt_apoe}
\title{Convert APOE Values}
\usage{
adt_apoe(
  apoecode,
  par_apoe = list(levels = c(3.4, 4.4, 2.4), labels = c(1, 2, NA))
)
}
\arguments{
\item{apoecode}{A string indicating the APOE. The available
options are: 3.4, 4.4, 2.4}

\item{par_apoe}{A list indicating the map for apoecode. Default
value is list(levels = c(3.4, 4.4, 2.4),labels = c(1, 2, NA))).}
}
\description{
Convert APOE values from (3.4, 4.4, 2.4) to (1, 2, NA).
}
