% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_get_dict}
\alias{adt_get_dict}
\title{Get dictionary}
\usage{
adt_get_dict(dict = c("par", "vars", "adni"), csv_fname = NULL)
}
\arguments{
\item{dict}{Name of the dictionary. The available options are:}

\item{csv_fname}{}
}
\description{
Get dictionary
}
\examples{
\dontrun{
dict_par <- adt_get_dict(dict = "par")
dict_par <- adt_get_dict(dict = "par", csv_fname = "foo_par.csv")}

}
