% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_nacc.R
\name{adt_get_nacc}
\alias{adt_get_nacc}
\title{Get NACC Data}
\usage{
adt_get_nacc(csv_fname)
}
\arguments{
\item{csv_fname}{CSV filename for the NACC data}
}
\value{
A dataframe with NACC data
}
\description{
Load NACC data from CSV files and convert the column names to be consistent
with other data sources
}
\examples{
\dontrun{
fname= "/Users/name/Documents/R/adbiomarker/Data/NACC"
dt_nacc <- get_nacc(path)}

}
