% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_query_cate}
\alias{adt_query_cate}
\title{Query for categorical variables}
\usage{
adt_query_cate(x, var)
}
\arguments{
\item{x}{The analysis data set}

\item{var}{A string indicating the queried variable}
}
\value{
A list with descriptive information and counts for each category.
}
\description{
Query for categorical variables
}
\examples{
\dontrun{
adt_query_cate(x, ""sex)
}
}
