% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\name{dict_adni}
\alias{dict_adni}
\title{ADNI Dictionary data}
\format{
A dataframe with the following variables:
\itemize{
  \item{file_label}{}
  \item{file_name}{}
}
}
\description{
Dictionary for ADNI files
}
