% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\docType{data}
\name{dict_par}
\alias{dict_par}
\title{Dictionary of columns}
\format{
A dataframe with the following variables:
\itemize{
  \item{file_code}{}
  \item{file_name}{}
}
}
\usage{
data(grav)
#'
}
\description{
Dictionary of columns
}
