% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\docType{data}
\name{dict_src_tables}
\alias{dict_src_tables}
\title{Dictionary of Column Names}
\format{

}
\usage{
data(dict_src_tables)
}
\description{
This dictionary maps the original column names (old_col_name) to the column
names of generated analysis dataset. In the raw data, the same variable may
have different names. This dictionary could be used to uniform variable
names. Also, as some column names in the raw data may be complicated
(including space, unit, special symbols, etc.), this dictionary is used to
simplify the column names.
}
