% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\name{dict_tbl}
\alias{dict_tbl}
\title{Dictionary of Table Settings}
\format{

}
\description{
This dictionary includes the basic setting of subtables, such as "file_code", "file_name", "start_row", and "date_format." 
In the raw data, the starting row and date type may vary for different subtables. This dictionary could be used to store these settings.
Also, since the subtables will have different names (maybe with time as a suffix), this dictionary is used to simplify the subtable names.
}
