% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\name{dict_vars}
\alias{dict_vars}
\title{Dictionary of variables}
\format{
A dataframe with the following variables:
\itemize{
  \item{file_code}{}
  \item{file_name}{}
}
}
\description{
Dictionary of variables
}
