% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_biocard.R
\name{get_biocard}
\alias{get_biocard}
\title{Get BIOCARD Data}
\usage{
get_biocard(
  path = "/Users/name/Data/BIOCARD",
  pattern = "*.xls",
  parameter_file = "par_file.xlsx",
  variable_file = "variables.xlsx"
)
}
\arguments{
\item{path}{Location where the biocard data stored}

\item{pattern}{Pattern of the data files}

\item{parameter_file}{Parameters dictionary (could be modified if needed)}

\item{variable_file}{Variables dictionary (could be modified if needed)}
}
\value{
a table with merged biocard data
}
\description{
Get BIOCARD Data
}
\examples{
\dontrun{
path = "/Users/name/Documents/R/adbiomarker/Data/BIOCARD"
dt_biocard <- get_biocard(path)
}

}
