% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_nacc.R
\name{get_nacc}
\alias{get_nacc}
\title{Get NACC Data}
\usage{
get_nacc(path)
}
\arguments{
\item{path}{The path where the nacc data stored}
}
\value{
a table with nacc data
}
\description{
Get NACC Data
}
\examples{
\dontrun{
path = "/Users/name/Documents/R/adbiomarker/Data/NACC"
dt_nacc <- get_nacc(path)
}
}
