% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{get_table}
\alias{get_table}
\title{Get Subtable}
\usage{
get_table(table_code, file_names, par_file, path)
}
\arguments{
\item{table_code}{Code of subtable ("Cognitive" as "COG")}

\item{file_names}{List of all files name in path}

\item{par_file}{Parameters dictionary (could be modified if needed)}

\item{path}{Location where data stored}
}
\value{

}
\description{
Get Subtable
}
\examples{
\dontrun{
dat_cog <- get_table("COG", file_names, par_file, path)
}

}
