% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_package.R
\name{parameters}
\alias{parameters}
\title{Parameters}
\arguments{
\item{src_type}{A string indicating the source of data. Options include
"BIOCARD", "NACC" and "ADNI"}

\item{src_col_name}{A string indicating a column name in the source data
file.}

\item{adt_tbl_code}{A string indicating the code of table, which can be mapped
to the source data files and is used for internal data manipulation.}

\item{dict}{A string indicating the name of the dictionary. The available
options are: "tbl": the dictionary for table structure, "col_name": the
dictionary for mapping column names, "adni": the dictionary for ADNI
data.}
}
\description{
List of function parameters in the \code{ADTool} package
}
