% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{plot.ad_ana_data}
\alias{plot.ad_ana_data}
\title{Plot for class ad_ana_data}
\usage{
\method{plot}{ad_ana_data}(ad_ana_data, ...)
}
\arguments{
\item{ad_ana_data}{}

\item{...}{}
}
\value{

}
\description{
Plot for class ad_ana_data
}
