% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{plot.adtool}
\alias{plot.adtool}
\title{Plot for class adtool}
\usage{
\method{plot}{adtool}(adtool, ...)
}
\arguments{
\item{adtool}{}

\item{...}{}
}
\value{

}
\description{
Plot for class adtool
}
