% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{summary.ad_ana_data}
\alias{summary.ad_ana_data}
\title{Summary for class ad_ana_data}
\usage{
\method{summary}{ad_ana_data}(dat, vars = NULL, ...)
}
\arguments{
\item{dat}{The baseline time dataset.}

\item{vars}{A list of strings indicating categorical variables name.}

\item{...}{}
}
\value{
The summary of this class
}
\description{
Summary for class ad_ana_data
}
