% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{summary.adtool}
\alias{summary.adtool}
\title{Summary for class adtool}
\usage{
\method{summary}{adtool}(adtool, ...)
}
\arguments{
\item{adtool}{}

\item{...}{}
}
\value{

}
\description{
Summary for class adtool
}
