% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_gen_tbl}
\alias{a_gen_tbl}
\title{Generate Information Table}
\usage{
a_gen_tbl(pat, group, distn)
}
\arguments{
\item{distn}{A string indicating the interested distribution. Options are "age", "visit", and "yar"}
}
\value{
A matrix with number of participants in different categories.
}
\description{
Generate Information Table
}
\examples{
\dontrun{
a_gen_tbl('sex', 'age')
}
}
