% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_map_var}
\alias{a_map_var}
\title{Get the column name in the source data file.}
\usage{
a_map_var(
  src_type = c("BIOCARD", "NACC", "ADNI"),
  tbl_code,
  col_name,
  dict_src_tables = NULL
)
}
\arguments{
\item{src_type}{A string indicating the source of data. Options include
"BIOCARD", "NACC" and "ADNI"}

\item{col_name}{A string indicating a column name in the source data file.}

\item{dict_src_tables}{A dictionary file for the parameters. The
default dictionary is appended in the package. Since the parameters in
tables may vary, it could be modified by using a costomized parameter
dictionary. The format can refer to the appended dictionary.
See \code{\link{dict_src_tables}} for more details.}
}
\value{
The real name in the source data file
}
\description{
Since the column names in tables may vary, it could be modified by using a
costomized column names dictionary. The format can refer to the appended
dictionary.
}
