% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_map_var}
\alias{a_map_var}
\title{Map Variables}
\usage{
a_map_var(data_name0 = "BIOCARD", table_code0, var_name, var_file)
}
\arguments{
\item{data_name0}{Dataset where the variable from}

\item{table_code0}{Table where the variable from}

\item{var_name}{Name of the variable}

\item{var_file}{Variables mapping file}
}
\value{
The real name in raw data
}
\description{
Map Variables
}
\examples{
\dontrun{
map_var("BIOCARD", "CSF", "date", var_file)
}


}
