% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_private.R
\name{a_read_file}
\alias{a_read_file}
\title{Read data from external files}
\usage{
a_read_file(code, file_names, dict_par)
}
\arguments{
\item{code}{Code of subtable ("Cognitive" as "COG")}

\item{file_names}{List of all files name in path}

\item{dict_par}{Parameters dictionary (could be modified if needed)}
}
\value{

}
\description{
Read data from external files
}
