% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_biocard.R
\name{adt_get_biocard}
\alias{adt_get_biocard}
\title{Get BIOCARD Data}
\usage{
adt_get_biocard(
  path = ".",
  pattern = "*.xls",
  dict_par = NULL,
  dict_vars = NULL,
  window = 730
)
}
\arguments{
\item{path}{Location where the biocard data stored}

\item{pattern}{Pattern of the data files}

\item{parameter_file}{Parameters dictionary (could be modified if needed)}

\item{variable_file}{Variables dictionary (could be modified if needed)}
}
\value{
A list with the following items
}
\description{
Load all BIOCARD data from a folder, generate the corresponding analysis
dataset with column names consistent with other data sources
}
\examples{
\dontrun{
dt_biocard <- get_biocard(path)
}

}
