% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_biocard.R
\name{adt_get_biocard}
\alias{adt_get_biocard}
\title{Import BIOCARD Data}
\usage{
adt_get_biocard(
  path = ".",
  merge_by = c("diagnosis", "cognitive", "csf", "hippocampus", "amydata", "entorhinal"),
  window = 730,
  window_overlap = FALSE,
  pattern = "*.xls",
  src_files = NULL,
  src_tables = NULL
)
}
\arguments{
\item{path}{Directory of the BIOCARD data files. Default value is the working
directory,}

\item{merge_by}{A character string indicating the source baseline time for
aligning the BIOCARD data when merging multiple files. Options include
"diagnosis", "cognitive", "csf", "hippocampus", "amydata" and
"entorhinal".}

\item{window}{An integer (unit of days) indicating the maximum acceptable gap
time for merging a biomarker test to base data. Default is 730 days. For
most data, the time windows are calculated from the baseline time (left
window = the midpoint between current and the previous time point; right
window = the midpoint between current and the next time point). If the
time windows are longer than the maximum acceptable length, then force to
select biomarker within the maximum window length we set. For the first
and last baseline time, since there is no "previous" or "next" time point
availiable, use the maximum acceptable window length instead.}

\item{window_overlap}{A logical value indicating the time window setting.
Default "False." If true, all time windows will set from 1/2 "window"
days before the baseline time to the 1/2 "window" days after baseline. In
this case, the time windows may overlap, which means some biomarkers may
be merged into multiple baseline data. If false, the windows will be
calculated from the blaseline times. The left window is set to be the
midpoint between the current and the previous time point. The right
window is set to be the midpoint between the current and the next time
point. For the first and last baseline time, since there is no "previous"
or "next" time point available, use the maximum acceptable window length
(set by the parameter "window") instead.}

\item{pattern}{A string indicating the pattern of all the data files. Default
is "*.xls" (should work for both .xls and .xlsx). This pattern is used to
read all table names from the path.}

\item{src_files}{Dictionary file for source file features. See
\link{\code{dict_src_files}} for more details.}

\item{src_tables}{Dictionary file for source table features. See
\link{\code{dict_src_tables}} for more details.}
}
\value{
The returned value is a list with the following items:

\itemize{

    \item data: The data includes patients' ids, baseline times,
    corresponding biomarkers, and biomarker test times.

    \item exid: The subject_ids of patients that are either not enrolled or
    impaired at baseline.
}
}
\description{
Import BIOCARD data from source files and generate the analysis dataset.
}
\examples{
\dontrun{
# with default unoverlaped window
dt_biocard <- get_biocard(path, merge_by = "dx")
# with costomized window
dt_biocard <- get_biocard(path, merge_by = "dx", window = 365,
                          window_overlap = TRUE)
# with costomized dictionary
dt_biocard <- get_biocard(path, merge_by = "dx",
                          dict_tbl = "\BIOCARD\DATA\dict_tbl.csv")
}

}
