% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_get_dict}
\alias{adt_get_dict}
\title{Get dictionary dataset}
\usage{
adt_get_dict(
  dict = c("src_files", "src_tables", "ana_data", "adni"),
  csv_fname = NULL
)
}
\arguments{
\item{dict}{A string indicating the name of the dictionary. The available
options are: "src_files": the dictionary for table structure, "src_tables": the
dictionary for mapping column names, "ana_data": the dictionary for query, "adni": the dictionary for ADNI
data.}

\item{csv_fname}{A string indicating the CSV filename to be saved (saved to
the default location). Default is NULL. The saved file is in .scv format.
If changes of default dictionaries needed, then the costomized dictionary
could be added here. However, the format of the costomized dictionary 
should follow the default one.}
}
\value{
The dictionary dataset.
}
\description{
Get the dictionary dataset as a dataframe or save as a csv file.
}
\examples{
\dontrun{
dict_par <- adt_get_dict(dict = "tbl")
dict_par <- adt_get_dict(dict = "tbl", csv_fname = "foo_tbl.csv")}

}
