% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_get_dict}
\alias{adt_get_dict}
\title{Get dictionary}
\usage{
adt_get_dict(dict = c("tbl", "col_name", "adni"), csv_fname = NULL)
}
\arguments{
\item{dict}{Name of the dictionary. 
The available options are: "tbl", "col_name", "adni"}

\item{csv_fname}{External dictionary. If change of default setting needed}
}
\description{
Get dictionary
}
\examples{
\dontrun{
dict_par <- adt_get_dict(dict = "tbl")
dict_par <- adt_get_dict(dict = "tbl", csv_fname = "foo_tbl.csv")}

}
