% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_get_dict}
\alias{adt_get_dict}
\title{Get dictionary}
\usage{
adt_get_dict(dict = c("tbl", "col_name", "adni"), csv_fname = NULL)
}
\arguments{
\item{dict}{A string indicating the name of the dictionary. 
The available options are: "tbl": the dictionary for table structure, 
"col_name": the dictionary for mapping column names, "adni": the dictionary for ADNI data.}

\item{csv_fname}{The path of external dictionary. If changes of default dictionaries needed, then the costomized dictionary could be added here. 
However, the format of the costomized dictionary should follow the default one.}
}
\value{
The dictionary dateset.
}
\description{
Get the needed dictionary data for leading subtables and mapping variables.
}
\examples{
\dontrun{
dict_par <- adt_get_dict(dict = "tbl")
dict_par <- adt_get_dict(dict = "tbl", csv_fname = "foo_tbl.csv")}

}
