% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_class_methods.R
\name{adt_query}
\alias{adt_query}
\title{Search description from dictionary}
\usage{
adt_query(biocard, var, opt = c("general", "variable", "sources"))
}
\arguments{
\item{biocard}{the object}

\item{var}{The variable name to query}

\item{opt}{Searching options}
}
\value{

}
\description{
Search description from dictionary
}
\examples{
\dontrun{
adt_query(biovard, "EDUC", "general")
}

}
