% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adt_toolkits.R
\name{adt_tk_query}
\alias{adt_tk_query}
\title{Query data frames}
\usage{
adt_tk_query(data, str = NULL, csv_fname = NULL, ignore.case = TRUE)
}
\arguments{
\item{data}{data sets}

\item{str}{String to search}

\item{csv_fname}{CSV filename to be saved}

\item{ignore.case}{Ignore case during matching}
}
\description{
Query data frames
}
