% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_functions.R
\name{plot.ad_ana_data}
\alias{plot.ad_ana_data}
\title{Plot for class ad_ana_data}
\usage{
\method{plot}{ad_ana_data}(
  dat,
  distn = c("age", "visit", "year"),
  group = c("diagnosis", "sex"),
  baseline = TRUE,
  ...
)
}
\arguments{
\item{dat}{The baseline time dataset.}

\item{distn}{A string indicating the interested distribution. Options are "age", "visit", and "yar"}

\item{...}{}
}
\value{

}
\description{
Plot for class ad_ana_data
}
